refactor(vla): consolidate validation tooling
Centralize LingBot VLA v2 action comparison and latency statistics, and move CUDA Graph coverage into its own model test module.

Remove the superseded in-process service benchmark and model-specific GPU capacity estimator while retaining upstream parity, runtime quantization timing, structured service validation, AIPerf, and the GPU release suite. Update documentation to reflect the latest H100 release-gate status.

Verification: ruff and formatting checks passed; 113 LingBot VLA v2 tests passed including CUDA Graph GPU replay; 9 AIPerf structured adapter tests passed; retained validation CLI entrypoints passed import/help smoke checks.
